sidecar/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0
pyyaml>=6.0.0

# Optional: Metrics
prometheus-client>=0.17.0

# Optional: gRPC
grpclib>=0.4.0
protobuf>=4.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
